requirements.txt
PyJWT==2.6.0
fastapi==0.109.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
pydantic==2.6.4
uvicorn==0.29.0
psycopg2-binary==2.9.9
bcrypt==4.1.1